requirements.txt
# Core dependencies
pyyaml>=6.0         # For configuration handling
paramiko>=2.7.2     # For SSH connections
python-nmap>=0.7.1  # For network scanning
pytz>=2023.3        # For timezone handling
requests>=2.28.1    # For API requests

# Optional dependencies
# libimobiledevice    # For iOS device interaction - system package 
# python-usblib>=1.0  # For USB device scanning
